perf(encoder): address re-review on j2c_dst_memory overflow safety
Two re-review findings from Copilot on PR #386:

- Widen `pos` from `uint32_t` to `size_t`.  After the raw-buffer change
  `pos` is used directly as an index into `buf`; a codestream >= 4 GiB
  (cinema-grade 16K HDR is in this territory) would wrap to a small
  index and cause out-of-bounds writes / memory corruption.

- `flush(std::ofstream&)` now casts to `std::streamsize` (not `long`)
  with a bounds check + dst.good() failure reporting.  On Windows LLP64
  `long` is 32-bit and would silently truncate codestreams >= 2 GiB
  without surfacing an error.

402/402 tests still pass.

+++ source/core/codestream/codestream.hpp
@@ -82,7 +82,10 @@ class j2c_dst_memory {
   // The raw buffer skips zero-init entirely; writes go straight to memory.
   uint8_t *buf;
   size_t   capacity;
-  uint32_t pos;
+  // pos is size_t so it cannot wrap for codestreams >= 4 GiB (uint32_t would
+  // overflow when used as an index into buf for very large outputs, e.g.,
+  // cinema-grade 16K HDR encodes).
+  size_t   pos;
   bool     is_flushed;
 
   // Grow capacity to at least `need` bytes.  Geometric growth (×2) to amortize
